Third Commit - fixed methods related to page objects and its derived classes
