+++ test/pageobjects/commonWebComponents/components/notification.js
@@ -3,6 +3,10 @@ const ComponentsUtility = require('../utils/componentsUtility.js');
 
 class Notification extends PageObject{
 
+    constructor(selector){
+        super(selector);
+    }
+
 }
 
 module.exports = Notification;
